+++ addon/components/flexberry-text-cell.js
@@ -13,6 +13,7 @@ export default FlexberryBaseComponent.extend({
 
     @property value
     @type String
+    @default undefined
   */
   value: undefined,
 
@@ -22,6 +23,7 @@ export default FlexberryBaseComponent.extend({
 
     @property maxTextLength
     @type Integer
+    @default 0
   */
   maxTextLength: 0,
 
@@ -30,6 +32,7 @@ export default FlexberryBaseComponent.extend({
 
     @property cutBySpaces
     @type Boolean
+    @default false
   */
   cutBySpaces: false,