+++ CHANGELOG.md
@@ -10,9 +10,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Add `flexberry-text-cell` for displaying limited text in olv component.
 * Blueprints:
     * Add setting offline database name.
+* Add `cut-string-by-length` util for cutting string by specified length.
 ### Fixed
 * `flexberry-objectlistview` component:
     * Fix paging and expand in hierarchical mode at mobile.
+### Changed
+* `object-list-view-cell` component:
+    * Component template was changed. If it's redefined in your project, you need to change it to support new features.
 
 ## [2.2.0-beta.14] - 2019-09-12]
 ### Added